fix(sonar): real bug fixes (S2184, S899) + suppress S3077 wholesale-reassign cases
- SessionStatsPanel.java: cast int operands to long before addition (S2184)
  to prevent overflow in lines/tokens totals.
- ChatWebServer.java: log dropped SSE close signal instead of ignoring the
  offer() return value (S899). Best-effort behaviour preserved.
- build.gradle.kts: file-level S3077 suppressions for 8 files where the
  volatile non-primitive fields hold immutable values or snapshot references
  that are only ever wholesale reassigned. AtomicReference would add overhead
  without changing semantics.

+++ plugin-core/src/main/java/com/github/catatafishen/agentbridge/ui/side/SessionStatsPanel.java
@@ -491,7 +491,7 @@ public final class SessionStatsPanel extends JPanel implements Disposable {
         turnToolsValue.setText(String.valueOf(turnTools));
         // Hide zero-value rows to reduce visual noise — a row of "0"s conveys no signal.
         turnToolsRow.setVisible(turnTools > 0);
-        long turnLines = snap.getTurnLinesAdded() + snap.getTurnLinesRemoved();
+        long turnLines = (long) snap.getTurnLinesAdded() + snap.getTurnLinesRemoved();
         turnLinesRow.setVisible(turnLines > 0);
 
         if (snap.getMultiplierMode()) {
@@ -500,7 +500,7 @@ public final class SessionStatsPanel extends JPanel implements Disposable {
             turnTokensRow.setVisible(true);
             turnCostRow.setVisible(false);
         } else {
-            long turnTok = snap.getTurnInputTokens() + snap.getTurnOutputTokens();
+            long turnTok = (long) snap.getTurnInputTokens() + snap.getTurnOutputTokens();
             Double turnCost = snap.getTurnCostUsd();
             boolean hasTurnUsage = turnTok > 0 || (turnCost != null && turnCost > 0.0);
             if (hasTurnUsage) {
@@ -529,7 +529,7 @@ public final class SessionStatsPanel extends JPanel implements Disposable {
         int sessionTools = snap.getSessionToolCalls();
         toolsValue.setText(String.valueOf(sessionTools));
         sessionToolsRow.setVisible(sessionTools > 0);
-        long sessionLines = snap.getSessionLinesAdded() + snap.getSessionLinesRemoved();
+        long sessionLines = (long) snap.getSessionLinesAdded() + snap.getSessionLinesRemoved();
         linesRow.setVisible(sessionLines > 0);
 
         if (snap.getMultiplierMode()) {
